fix: import promps async

+++ src/utils.ts
@@ -3,6 +3,27 @@ import { isPackageExists } from 'local-pkg'
 import c from 'picocolors'
 import type { Awaitable, TypedFlatConfigItem } from './types'
 
+export const parserPlain = {
+    meta: {
+        name: 'parser-plain',
+    },
+    parseForESLint: (code: string) => ({
+        ast: {
+            body: [],
+            comments: [],
+            loc: { end: code.length, start: 0 },
+            range: [0, code.length],
+            tokens: [],
+            type: 'Program',
+        },
+        scopeManager: null,
+        services: { isPlain: true },
+        visitorKeys: {
+            Program: [],
+        },
+    }),
+}
+
 /**
  * Combine array and non-array configs into a single array.
  */
@@ -84,44 +105,18 @@ export async function interopDefault<T>(m: Awaitable<T>): Promise<T extends { de
     return (resolved as any).default || resolved
 }
 
-export async function ensurePackages(packages: string[]) {
+export async function ensurePackages(packages: (string | undefined)[]) {
     if (process.env.CI || process.stdout.isTTY === false)
         return
 
-    const nonExistingPackages = packages.filter(i => !isPackageExists(i))
+    const nonExistingPackages = packages.filter(i => i && !isPackageExists(i)) as string[]
     if (nonExistingPackages.length === 0)
         return
 
-    const { default: prompts } = await import('prompts')
-    const { result } = await prompts([
-        {
-            initial: true,
-            message: `The following ${nonExistingPackages.length === 1 ? 'package is' : 'packages are'} required for this config: ${c.green(nonExistingPackages.join(c.white(', ')))}. Do you want to install them?`,
-            name: 'result',
-            type: 'confirm',
-        },
-    ])
+    const p = await import('@clack/prompts')
+    const result = await p.confirm({
+        message: `The following ${nonExistingPackages.length === 1 ? 'package is' : 'packages are'} required for this config: ${c.green(nonExistingPackages.join(c.white(', ')))}. Do you want to install them?`,
+    })
     if (result)
         await import('@antfu/install-pkg').then(i => i.installPackage(nonExistingPackages, { dev: true }))
-}
-
-export const parserPlain = {
-    meta: {
-        name: 'parser-plain',
-    },
-    parseForESLint: (code: string) => ({
-        ast: {
-            body: [],
-            comments: [],
-            loc: { end: code.length, start: 0 },
-            range: [0, code.length],
-            tokens: [],
-            type: 'Program',
-        },
-        scopeManager: null,
-        services: { isPlain: true },
-        visitorKeys: {
-            Program: [],
-        },
-    }),
 }
